dependencies loaded first.
There could be scenarios where Tinymce appears before Wysiwyg in Hook.bootstraps. But Wysiwyg is a dependency plugin of Tinymce. In that case we need to reorder the list of plugins to be loaded.

This could become a part of a lib in the future or done in Extensions plugin whenever a plugin is activated/deactivated.