Fix: Populate 'created' field correctly. Closes #355
In order for 'created' field to be populated correctly by cake, the
key itself must not be present at all during in the POSTed data.

Unsetting 'created' in beforeSave() will not work since the callback is
dispatched _only after_ the date population logic is executed in Model::save().
By that time, the 'created' value is already set to '' which will passed to
the Dbo layer, causing invalid date '0000-00-00 00:00:00' in the table.

This change solves the problem by copying the date population behavior from
Model::save(). When, empty 'created' field is POSTed, the value is then
recalculated with the current datetime. User supplied 'created' value is
also accepted and not clobbered.