Update navbar.less
Added to .sidebar .sidebar-item span (lines 46-50). A cross browser method (as per the original post by the [Mozilla Team](http://blog.mozilla.org/webdev/2009/02/20/cross-browser-inline-block/)) of displaying the span inline-block.

This means that when a menu item double lines (for instance add a 
tag to a title in your CroogoNav::add() declaration) the text lines up neatly.

This pull request is in response to @rchavik requesting the css to go into the navbar.less file rather than the croogo-boostrap.css file.