Add method: output rendered block
The ability to call a block from with a block or node added in v2 is great, but it only replaces the block shortcode with the body field?  What if the title needs displaying or the block uses a custom element? 

https://github.com/croogo/croogo/commit/b131edc53bebe27422f5e437f6301541bd219b70

I propose a new RegionsHelper method that uses similar logic to the blocks() method to take data from blocks_for_layout and output a single rendered block.  This can then be called from places like Blocks  EventHandler::filterBlockShortcode().