Added core plugins to settings seed
I was playing around with Croogo 3.x for a little and tried to install my own extension. Since the .zip upload seems kind of broken I just put it in the plugin folder. It then showed up on the plugins list, but I couldn't activate it because somehow Croogo was convinced that Croogo/Extensions (and Croogo/Acl for that matter) were not activated. However there was no option to activate them, so I guess they should be activated by default which makes this the right place to put them, I hope :)